Revert ECONNRESET observability suppression from #59287 (#60399)
