+++ src/observability/middleware/handle-errors.ts
@@ -1,7 +1,6 @@
 import type { NextFunction, Response } from 'express'
 
 import FailBot from '../lib/failbot'
-import { shouldLogException, type ErrorWithCode } from '../lib/should-log-exception'
 import { nextApp } from '@/frame/middleware/next'
 import { minimumNotFoundHtml } from '@/frame/lib/constants'
 import { setFastlySurrogateKey, SURROGATE_ENUMS } from '@/frame/middleware/set-fastly-surrogate-key'
@@ -13,6 +12,26 @@ import { createLogger } from '@/observability/logger'
 const logger = createLogger(import.meta.url)
 
 const DEBUG_MIDDLEWARE_TESTS = Boolean(JSON.parse(process.env.DEBUG_MIDDLEWARE_TESTS || 'false'))
+
+type ErrorWithCode = Error & {
+  code: string
+  statusCode?: number
+  status?: string
+}
+
+function shouldLogException(error: ErrorWithCode) {
+  const IGNORED_ERRORS = [
+    // Client connection aborted
+    'ECONNRESET',
+  ]
+
+  if (IGNORED_ERRORS.includes(error.code)) {
+    return false
+  }
+
+  // We should log this exception
+  return true
+}
 
 async function logException(error: ErrorWithCode, req: ExtendedRequest) {
   if (process.env.NODE_ENV !== 'test' && shouldLogException(error)) {
